Use official FlutterAppLifecycle callback on web for Flutter 3.22 onwards
